Move exemplar SPI interface from api.spi to spi
