+++ metrics/api/src/main/java/io/helidon/metrics/api/ExemplarServiceManager.java
@@ -23,10 +23,10 @@ import java.util.function.Predicate;
 import java.util.function.Supplier;
 
 import io.helidon.common.HelidonServiceLoader;
-import io.helidon.metrics.api.spi.ExemplarService;
+import io.helidon.metrics.spi.ExemplarService;
 
 /**
- * Loads the {@link io.helidon.metrics.api.spi.ExemplarService} instance (if any) with the most urgent priority.
+ * Loads the {@link io.helidon.metrics.spi.ExemplarService} instance (if any) with the most urgent priority.
  */
 class ExemplarServiceManager {
