+++ metrics/api/src/main/java/io/helidon/metrics/spi/ExemplarService.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2021 Oracle and/or its affiliates.
+ * Copyright (c) 2021, 2023 Oracle and/or its affiliates.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,7 +13,16 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+package io.helidon.metrics.spi;
+
 /**
- * SPI for the metrics API.
+ * Behavior for supporting exemplars in metrics histograms.
  */
-package io.helidon.metrics.api.spi;
+public interface ExemplarService {
+    /**
+     * Returns a label using whatever current context is available.
+     *
+     * @return the label
+     */
+    String label();
+}
